+++ CHANGELOG.md
@@ -104,6 +104,7 @@
 - MOVE : add a control on period status
 - BANK ORDER : add a viewer with technical origin (Automatic,Entry,Import) information display.
 - BANK ORDER : when technical origin is automatic, it's not possible to edit the bank order.
+- DEBT RECOVERY : send email when manualValidationOK = true for DebtRecoveryMethodLine.
 
 ## Bug Fixes
 - PURCHASE ORDER : Fix error when merging a set of Purchase Orders with no trading names